# openapi.yaml
openapi: 3.0.3
info:
  title: Sacsol API
  version: 1.0.0
  description: Inventory management API for Sacsol
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /api/grn/:
    get:
      operationId: grn_list
      summary: List goods receipts
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - Procurement / GRN
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGoodsReceiptList'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      operationId: grn_create
      summary: Create goods receipt
      tags:
      - Procurement / GRN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodsReceipt'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/grn/{id}/:
    get:
      operationId: grn_retrieve
      summary: Get goods receipt
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this goods receipt.
        required: true
      tags:
      - Procurement / GRN
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodsReceipt'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      operationId: grn_update
      summary: Replace goods receipt
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this goods receipt.
        required: true
      tags:
      - Procurement / GRN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GoodsReceiptRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodsReceipt'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    patch:
      operationId: grn_partial_update
      summary: Update goods receipt
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this goods receipt.
        required: true
      tags:
      - Procurement / GRN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedGoodsReceiptRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedGoodsReceiptRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedGoodsReceiptRequest'
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodsReceipt'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      operationId: grn_destroy
      summary: Delete goods receipt
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this goods receipt.
        required: true
      tags:
      - Procurement / GRN
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/lpos/:
    get:
      operationId: lpo_list
      summary: List LPOs
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - Procurement / LPO
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLPOList'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      operationId: lpo_create
      summary: Create LPO
      tags:
      - Procurement / LPO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LPORequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LPORequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LPORequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LPO'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/lpos/{id}/:
    get:
      operationId: lpo_retrieve
      summary: Get LPO
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this lpo.
        required: true
      tags:
      - Procurement / LPO
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LPO'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      operationId: lpo_update
      summary: Replace LPO
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this lpo.
        required: true
      tags:
      - Procurement / LPO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LPORequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LPORequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LPORequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LPO'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    patch:
      operationId: lpo_partial_update
      summary: Update LPO
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this lpo.
        required: true
      tags:
      - Procurement / LPO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedLPORequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedLPORequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedLPORequest'
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LPO'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      operationId: lpo_destroy
      summary: Delete LPO
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this lpo.
        required: true
      tags:
      - Procurement / LPO
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/lpos/{id}/attachments/{att_id}/:
    delete:
      operationId: lpo_delete_attachment
      summary: Delete LPO attachment
      parameters:
      - in: path
        name: att_id
        schema:
          type: string
          pattern: ^\d+$
        required: true
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this lpo.
        required: true
      tags:
      - Procurement / LPO
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
components:
  schemas:
...
    GoodsReceipt:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        lpo:
          type: integer
        received_by:
          type: integer
          readOnly: true
        received_at:
          type: string
          format: date-time
          readOnly: true
        reference:
          type: string
          maxLength: 64
        note:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/GRNItemIn'
      required:
      - items
      - lpo
    GoodsReceiptRequest:
      type: object
      properties:
        lpo:
          type: integer
        reference:
          type: string
          maxLength: 64
        note:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/GRNItemInRequest'
      required:
      - items
      - lpo
...
    PaginatedGoodsReceiptList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/GoodsReceipt'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: /